Move OctreeLevelInfo and other cleanup/fixes
